+++ src/preprocess/transforms/query.test.ts
@@ -75,8 +75,7 @@ describe('query preprocessor', function () {
 		    queryHandler: _TestQuery_handler,
 		    config: houdiniConfig,
 		    artifact: _TestQueryArtifact,
-		    variableFunction: null,
-		    getProps: () => $$props
+		    variableFunction: null
 		});
 
 		$:
@@ -201,8 +200,7 @@ describe('query preprocessor', function () {
 		    queryHandler: _TestQuery1_handler,
 		    config: houdiniConfig,
 		    artifact: _TestQuery1Artifact,
-		    variableFunction: null,
-		    getProps: () => $$props
+		    variableFunction: null
 		});
 
 		const {
@@ -211,8 +209,7 @@ describe('query preprocessor', function () {
 		    queryHandler: _TestQuery2_handler,
 		    config: houdiniConfig,
 		    artifact: _TestQuery2Artifact,
-		    variableFunction: null,
-		    getProps: () => $$props
+		    variableFunction: null
 		});
 
 		$:
@@ -323,8 +320,7 @@ describe('query preprocessor', function () {
 		    queryHandler: _TestQuery_handler,
 		    config: houdiniConfig,
 		    artifact: _TestQueryArtifact,
-		    variableFunction: TestQueryVariables,
-		    getProps: () => $$props
+		    variableFunction: TestQueryVariables
 		});
 
 		$:
@@ -406,8 +402,7 @@ describe('query preprocessor', function () {
 		    queryHandler: _TestQuery_handler,
 		    config: houdiniConfig,
 		    artifact: _TestQueryArtifact,
-		    variableFunction: null,
-		    getProps: () => $$props
+		    variableFunction: null
 		});
 
 		$:
@@ -612,8 +607,7 @@ describe('query preprocessor', function () {
 		    queryHandler: _TestQuery_handler,
 		    config: houdiniConfig,
 		    artifact: _TestQueryArtifact,
-		    variableFunction: TestQueryVariables,
-		    getProps: () => $$props
+		    variableFunction: TestQueryVariables
 		});
 
 		$:
@@ -1003,6 +997,10 @@ test('afterLoad hook', async function () {
 		        "framework": "sapper",
 		        "hookFn": afterLoad,
 
+		        "input": {
+		            "TestQuery": _TestQuery_Input
+		        },
+
 		        "data": {
 		            "TestQuery": _TestQuery.result.data
 		        }
@@ -1135,6 +1133,11 @@ test('afterLoad hook - multiple queries', async function () {
 		        "variant": "after",
 		        "framework": "sapper",
 		        "hookFn": afterLoad,
+
+		        "input": {
+		            "TestQuery1": _TestQuery1_Input,
+		            "TestQuery2": _TestQuery2_Input
+		        },
 
 		        "data": {
 		            "TestQuery1": _TestQuery1.result.data,
@@ -1283,6 +1286,10 @@ test('both beforeLoad and afterLoad hooks', async function () {
 		        "variant": "after",
 		        "framework": "sapper",
 		        "hookFn": afterLoad,
+
+		        "input": {
+		            "TestQuery": _TestQuery_Input
+		        },
 
 		        "data": {
 		            "TestQuery": _TestQuery.result.data
@@ -1515,8 +1522,7 @@ test('2 queries, one paginated one not', async function () {
 		    queryHandler: _TestQuery1_handler,
 		    config: houdiniConfig,
 		    artifact: _TestQuery1Artifact,
-		    variableFunction: TestQuery1Variables,
-		    getProps: () => $$props
+		    variableFunction: TestQuery1Variables
 		});
 
 		const {
@@ -1525,8 +1531,7 @@ test('2 queries, one paginated one not', async function () {
 		    queryHandler: _TestQuery2_handler,
 		    config: houdiniConfig,
 		    artifact: _TestQuery2Artifact,
-		    variableFunction: TestQuery2Variables,
-		    getProps: () => $$props
+		    variableFunction: TestQuery2Variables
 		});
 
 		$: